Revert "Cipher dev"
